refactor: flatter structure
