+++ src/psycop_model_evaluation/binary_classification/global_performance/roc_auc.py
