--- src/psycop_model_evaluation/binary_classification/tables/__init__.py
@@ -1,1 +1,0 @@
-
